COMPONENTS
In the CMakeLists files, support for project components (own libraries) has been added, and the initialization functions have been replaced with the new init_fun libraries
