+++ CMakeLists.txt
@@ -8,8 +8,9 @@ cmake_minimum_required(VERSION 3.5)
 # Dołączenie pliku CMake z SDK (ten plik ustawia środowisko ESP-IDF)
 include($ENV{IDF_PATH}/tools/cmake/project.cmake)
 
+# Ustawienie zmiennej środowiskowej z ścieżką do SDK i własnych komponentów (wewnątrz kontenera)
+set(EXTRA_COMPONENT_DIRS /opt/esp/ESP8266_RTOS_SDK/components
+    "${PROJECT_DIR}/../components")
+
 # Nazwa projektu
 project(nESP_WiFi_Tutorials)
-
-# Ustawienie zmiennej środowiskowej z ścieżką do SDK (wewnątrz kontenera)
-set(EXTRA_COMPONENT_DIRS /opt/esp/ESP8266_RTOS_SDK/components)
