+++ components/init_fun/CMakeLists.txt
@@ -1,0 +1,4 @@
+idf_component_register(SRCS "init_fun.c"
+                       INCLUDE_DIRS "."
+                       PRIV_INCLUDE_DIRS "."
+                       REQUIRES main) # Dodanie komponentów nagłówka pliku źródłowego
