+++ main/CMakeLists.txt
@@ -1,1 +1,2 @@
-idf_component_register(SRCS "main.c" INCLUDE_DIRS ".")
+idf_component_register(SRCS "main.c"
+                       INCLUDE_DIRS ".")
