+++ main/main.h
@@ -1,0 +1,16 @@
+#ifndef __MAIN_H__
+#define __MAIN_H__
+
+#include <FreeRTOS.h>
+#include <task.h>
+#include <stdio.h>
+#include <string.h>
+
+#include <gpio.h>
+#include <esp_log.h>
+#include <esp_event.h>
+#include <esp_netif.h>
+#include <esp_wifi.h>
+#include <nvs_flash.h>
+
+#endif
